+++ src/config/dependencyConfigESS.ts
@@ -23,6 +23,7 @@ import PostgresTemplateDataSource from '../datasources/postgres/TemplateDataSour
 import PostgresUnitDataSource from '../datasources/postgres/UnitDataSource';
 import PostgresUserDataSource from '../datasources/postgres/UserDataSource';
 import PostgresVisitDataSource from '../datasources/postgres/VisitDataSource';
+import { essEmailHandler } from '../eventHandlers/email/essEmailHandler';
 import { SparkPostMailService } from '../eventHandlers/MailService/SparkPostMailService';
 import {
   createListenToRabbitMQHandler,
@@ -64,6 +65,8 @@ mapClass(Tokens.UserAuthorization, UserAuthorization);
 mapClass(Tokens.AssetRegistrar, EAMAssetRegistrar);
 
 mapClass(Tokens.MailService, SparkPostMailService);
+
+mapValue(Tokens.EmailEventHandler, essEmailHandler);
 
 mapValue(Tokens.PostToMessageQueue, createPostToRabbitMQHandler());
 mapValue(Tokens.ListenToMessageQueue, createListenToRabbitMQHandler());
