+++ src/eventHandlers/email/stfcEmailHandler.ts
@@ -1,0 +1,78 @@
+import { logger } from '@user-office-software/duo-logger';
+import { container } from 'tsyringe';
+
+import { Tokens } from '../../config/Tokens';
+import { UserDataSource } from '../../datasources/UserDataSource';
+import { ApplicationEvent } from '../../events/applicationEvents';
+import { Event } from '../../events/event.enum';
+import EmailSettings from '../MailService/EmailSettings';
+import { MailService } from '../MailService/MailService';
+
+export async function stfcEmailHandler(event: ApplicationEvent) {
+  const mailService = container.resolve<MailService>(Tokens.MailService);
+  const userDataSource = container.resolve<UserDataSource>(
+    Tokens.UserDataSource
+  );
+
+  if (event.isRejection) {
+    return;
+  }
+
+  switch (event.type) {
+    case Event.PROPOSAL_SUBMITTED: {
+      const principalInvestigator = await userDataSource.getUser(
+        event.proposal.proposerId
+      );
+      const participants = await userDataSource.getProposalUsersFull(
+        event.proposal.primaryKey
+      );
+      if (!principalInvestigator) {
+        return;
+      }
+
+      const options: EmailSettings = {
+        content: {
+          template_id: 'proposal-submitted',
+        },
+        substitution_data: {
+          piPreferredname: principalInvestigator.preferredname,
+          piLastname: principalInvestigator.lastname,
+          proposalNumber: event.proposal.proposalId,
+          proposalTitle: event.proposal.title,
+          coProposers: participants.map(
+            (partipant) => `${partipant.preferredname} ${partipant.lastname} `
+          ),
+          call: '',
+        },
+        recipients: [
+          { address: principalInvestigator.email },
+          ...participants.map((partipant) => {
+            return {
+              address: {
+                email: partipant.email,
+                header_to: principalInvestigator.email,
+              },
+            };
+          }),
+        ],
+      };
+
+      mailService
+        .sendMail(options)
+        .then((res: any) => {
+          logger.logInfo('Emails sent on proposal submission:', {
+            result: res,
+            event,
+          });
+        })
+        .catch((err: string) => {
+          logger.logError('Could not send email(s) on proposal submission:', {
+            error: err,
+            event,
+          });
+        });
+
+      return;
+    }
+  }
+}
