+++ src/eventHandlers/index.ts
@@ -1,13 +1,19 @@
+import { container } from 'tsyringe';
+
+import { Tokens } from '../config/Tokens';
 import { ApplicationEvent } from '../events/applicationEvents';
 import { EventHandler } from '../events/eventBus';
-import createEmailHandler from './email';
 import createLoggingHandler from './logging';
 import { createPostToQueueHandler } from './messageBroker';
 import createProposalWorkflowHandler from './proposalWorkflow';
 
 export default function createEventHandlers(): EventHandler<ApplicationEvent>[] {
+  const emailHandler = container.resolve<
+    (event: ApplicationEvent) => Promise<void>
+  >(Tokens.EmailEventHandler);
+
   return [
-    createEmailHandler(),
+    emailHandler,
     createLoggingHandler(),
     createPostToQueueHandler(),
     createProposalWorkflowHandler(),
